Add build.gradle
Presently, we require a full-blown eclipse IDE in order to compile the
AddressBook and run tests.

However, in the future we want to implement Travis CI support, which
requires a smaller build system that can run be run on the command line
on headless systems.

Gradle is one such build system. It is chosen because it seems that
AddressBook-Level4 will also be using it, so it's good to get familiar
with it.

Add a build.gradle file which contains build instructions for gradle. In
particular, it instructs gradle that:

* build products will be output in "bin"

* main application sources and resources are in src/

* test sources are in test/java/

* automatically download junit 4.12 from maven central for running tests

Users can build the application by running:

    ./gradlew build

and run tests with:

    ./gradlew check