Convert eclipse project to gradle project
There are now two ways to build the address book: with eclipse and with
gradle. This is bad as it is not DRY.

Make the gradle build system the authoritative one by converting the
eclipse project into a gradle project. This will activate gradle
integration in eclipse, so we can build the project using gradle from
the comfort of the eclipse IDE.