Add .travis.yml
Travis CI is a continuous integration service that will automatically
run tests on all branches and PRs in the repository to ensure that code
that breaks tests will not be merged into master.

Add a .travis.yml file which will tell Travis CI how to compile and run
the tests of the AddressBook. This was derived from the Travis CI
guide[1].

In particular, we tell Travis CI to build and run the tests with Oracle
JDK 8 (Since we depend on Java 8 features).

Travis CI will automatically detect that we are using gradle, and will
run "gradle check".

[1] https://docs.travis-ci.com/user/languages/java