Remove all eclipse project files
bda0b3c (Convert eclipse project to gradle project, 2016-09-16)
attempted to convert the existing Eclipse project files in the
repository to become an Eclipse gradle project. However, it turns out
that it is not effective, as Eclipse may throw the following error when
importing the project with the "General->Existing Projects into
Workspace" importer:

    Errors running builder 'Java Builder' on project 'AddressBook-Level3':
    Cannot read preference from file

Technically, this is due to /.settings/org.eclipse.jdt.core.prefs file
being missing, which is in turn caused by the /.settings directory
being mistakenly ignored.

However, let's just bite the bullet and do a clean removal of all the
eclipse project files in this repository. With this change, developers
can no longer import this project with "General->Existing Projects into
Workspace" importer, as there is no eclipse project file for them to
import. Instead, they have to use the "Gradle->Gradle Project" importer.

This is not such a bad thing -- the Gradle project importer is superior
to the "General->Existing Projects into Workspace" importer as it will
also download dependencies, while the latter will not.

Update the .gitignore and developer guide to reflect this change.